Solution "Cat" for theme oop-principles.
