+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/cat/Cat.java
@@ -1,9 +1,9 @@
-package com.rakovets.course.java.core.practice.oop.principles;
+package com.rakovets.course.java.core.practice.oop.principles.cat;
 
 public class Cat {
     private String name;
 
-    Cat(String name) {
+    public Cat(String name) {
         this.name = name;
     }
