+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/cat/Person.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop.principles;
+package com.rakovets.course.java.core.practice.oop.principles.cat;
 
 public class Person {
     private int happiness;
