+++ src/test/java/com/rakovets/course/java/core/practice/oop/principles/CatDemo.java
@@ -1,4 +1,5 @@
 package com.rakovets.course.java.core.practice.oop.principles;
+import com.rakovets.course.java.core.practice.oop.principles.cat.*;
 
 public class CatDemo {
     public static void main(String[] args) {
